pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scaffold"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncmy>=0.2",
    "pymysql>=1.1",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
messaging = [
    "aio-pika>=9.4",
]
ai = [
    "httpx>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.9",
    "ruff>=0.4",
    "aio-pika>=9.4",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/scaffold"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
strict = true
